Add Described FlagName and RepoName instances
Move few Arbitrary instances to more correct places,
add tests
